+++ activerecord/test/cases/encryption/configurable_test.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require "cases/encryption/helper"
-require "models/book"
+require "models/book_encrypted"
 
 class ActiveRecord::Encryption::ConfigurableTest < ActiveRecord::TestCase
   test "can access context properties with top level getters" do
