+++ activerecord/test/cases/encryption/encryptable_record_api_test.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
 require "cases/encryption/helper"
-require "models/author"
-require "models/book"
-require "models/post"
+require "models/author_encrypted"
+require "models/book_encrypted"
+require "models/post_encrypted"
 
 class ActiveRecord::Encryption::EncryptableRecordApiTest < ActiveRecord::TestCase
   fixtures :posts
