+++ activerecord/test/cases/encryption/encrypted_fixtures_test.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require "cases/encryption/helper"
-require "models/book"
+require "models/book_encrypted"
 
 class ActiveRecord::Encryption::EncryptableFixtureTest < ActiveRecord::TestCase
   fixtures :encrypted_books, :encrypted_book_that_ignores_cases
