+++ activerecord/test/cases/encryption/performance/envelope_encryption_performance_test.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require "cases/encryption/helper"
-require "models/book"
+require "models/book_encrypted"
 
 class ActiveRecord::Encryption::EvenlopeEncryptionPerformanceTest < ActiveRecord::TestCase
   fixtures :encrypted_books
