+++ activerecord/test/models/book.rb
@@ -33,21 +33,3 @@ class PublishedBook < ActiveRecord::Base
 
   validates_uniqueness_of :isbn
 end
-
-class EncryptedBook < Book
-  self.table_name = "books"
-
-  encrypts :name, deterministic: true
-end
-
-class EncryptedBookWithDowncaseName < Book
-  self.table_name = "books"
-
-  encrypts :name, deterministic: true, downcase: true
-end
-
-class EncryptedBookThatIgnoresCase < Book
-  self.table_name = "books"
-
-  encrypts :name, deterministic: true, ignore_case: true
-end
