+++ activerecord/test/models/traffic_light.rb
@@ -4,7 +4,3 @@ class TrafficLight < ActiveRecord::Base
   serialize :state, Array
   serialize :long_state, Array
 end
-
-class EncryptedTrafficLight < TrafficLight
-  encrypts :state
-end
