+++ activerecord/test/models/traffic_light_encrypted.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+
+require "models/traffic_light"
+
+class EncryptedTrafficLight < TrafficLight
+  encrypts :state
+end
